+++ src/devices/cpu/sh/sh2.cpp
@@ -856,42 +856,7 @@ extern unsigned DasmSH2(std::ostream &stream, unsigned pc, uint16_t opcode);
 
 using namespace uml;
 
-/***************************************************************************
-    DEBUGGING
-***************************************************************************/
 
-#define SET_EA                      (0) // makes slower but "shows work" in the EA fake register like the interpreter
-
-#define ADDSUBV_DIRECT              (0)
-
-#if SET_EA
-#define SETEA(x) UML_MOV(block, mem(&m_sh2_state->ea), ireg(x))
-#else
-#define SETEA(x)
-#endif
-
-/***************************************************************************
-    CONSTANTS
-***************************************************************************/
-
-/* map variables */
-#define MAPVAR_PC                   M0
-#define MAPVAR_CYCLES                   M1
-
-/* exit codes */
-#define EXECUTE_OUT_OF_CYCLES           0
-#define EXECUTE_MISSING_CODE            1
-#define EXECUTE_UNMAPPED_CODE           2
-#define EXECUTE_RESET_CACHE         3
-
-#define PROBE_ADDRESS                   ~0
-
-
-/***************************************************************************
-    MACROS
-***************************************************************************/
-
-#define R32(reg)        m_regmap[reg]
 
 /***************************************************************************
     INLINE FUNCTIONS
@@ -1019,18 +984,6 @@ void sh2_device::func_unimplemented()
 	m_sh2_state->pending_nmi = 1;
 }
 
-/*-------------------------------------------------
-    cfunc_fastirq - checks for pending IRQs
--------------------------------------------------*/
-static void cfunc_fastirq(void *param)
-{
-	((sh2_device *)param)->func_fastirq();
-}
-
-void sh2_device::func_fastirq()
-{
-	sh2_exception("fastirq",m_sh2_state->irqline);
-}
 
 /*-------------------------------------------------
     cfunc_MAC_W - implementation of MAC_W Rm,Rn
@@ -1796,83 +1749,6 @@ void sh2_device::log_add_disasm_comment(drcuml_block *block, uint32_t pc, uint32
 	}
 }
 
-/*-------------------------------------------------
-    generate_update_cycles - generate code to
-    subtract cycles from the icount and generate
-    an exception if out
--------------------------------------------------*/
-void sh2_device::generate_update_cycles(drcuml_block *block, compiler_state *compiler, uml::parameter param, bool allow_exception)
-{
-	/* check full interrupts if pending */
-	if (compiler->checkints)
-	{
-		code_label skip = compiler->labelnum++;
-
-		compiler->checkints = false;
-		compiler->labelnum += 4;
-
-		/* check for interrupts */
-		UML_MOV(block, mem(&m_sh2_state->irqline), 0xffffffff);     // mov irqline, #-1
-		UML_CMP(block, mem(&m_sh2_state->pending_nmi), 0);          // cmp pending_nmi, #0
-		UML_JMPc(block, COND_Z, skip+2);                    // jz skip+2
-
-		UML_MOV(block, mem(&m_sh2_state->pending_nmi), 0);          // zap pending_nmi
-		UML_JMP(block, skip+1);                     // and then go take it (evec is already set)
-
-		UML_LABEL(block, skip+2);                   // skip+2:
-		UML_MOV(block, mem(&m_sh2_state->evec), 0xffffffff);        // mov evec, -1
-		UML_MOV(block, I0, 0xffffffff);         // mov r0, -1 (r0 = irq)
-		UML_AND(block, I1,  I0, 0xffff);                // and r1, r0, 0xffff
-
-		UML_LZCNT(block, I1, mem(&m_sh2_state->pending_irq));       // lzcnt r1, pending_irq
-		UML_CMP(block, I1, 32);             // cmp r1, #32
-		UML_JMPc(block, COND_Z, skip+4);                    // jz skip+4
-
-		UML_SUB(block, mem(&m_sh2_state->irqline), 31, I1);     // sub irqline, #31, r1
-
-		UML_LABEL(block, skip+4);                   // skip+4:
-		UML_CMP(block, mem(&m_sh2_state->internal_irq_level), 0xffffffff);  // cmp internal_irq_level, #-1
-		UML_JMPc(block, COND_Z, skip+3);                    // jz skip+3
-		UML_CMP(block, mem(&m_sh2_state->internal_irq_level), mem(&m_sh2_state->irqline));      // cmp internal_irq_level, irqline
-		UML_JMPc(block, COND_LE, skip+3);                   // jle skip+3
-
-		UML_MOV(block, mem(&m_sh2_state->irqline), mem(&m_sh2_state->internal_irq_level));      // mov r0, internal_irq_level
-
-		UML_LABEL(block, skip+3);                   // skip+3:
-		UML_CMP(block, mem(&m_sh2_state->irqline), 0xffffffff);     // cmp irqline, #-1
-		UML_JMPc(block, COND_Z, skip+1);                    // jz skip+1
-		UML_CALLC(block, cfunc_fastirq, this);               // callc fastirq
-
-		UML_LABEL(block, skip+1);                   // skip+1:
-		UML_CMP(block, mem(&m_sh2_state->evec), 0xffffffff);        // cmp evec, 0xffffffff
-		UML_JMPc(block, COND_Z, skip);                  // jz skip
-
-		UML_SUB(block, R32(15), R32(15), 4);            // sub R15, R15, #4
-		UML_MOV(block, I0, R32(15));                // mov r0, R15
-		UML_MOV(block, I1, mem(&m_sh2_state->irqsr));           // mov r1, irqsr
-		UML_CALLH(block, *m_write32);                    // call write32
-
-		UML_SUB(block, R32(15), R32(15), 4);            // sub R15, R15, #4
-		UML_MOV(block, I0, R32(15));                // mov r0, R15
-		UML_MOV(block, I1, param);              // mov r1, nextpc
-		UML_CALLH(block, *m_write32);                    // call write32
-
-		UML_HASHJMP(block, 0, mem(&m_sh2_state->evec), *m_nocode);       // hashjmp m_sh2_state->evec
-
-		UML_LABEL(block, skip);                         // skip:
-	}
-
-	/* account for cycles */
-	if (compiler->cycles > 0)
-	{
-		UML_SUB(block, mem(&m_sh2_state->icount), mem(&m_sh2_state->icount), MAPVAR_CYCLES);    // sub     icount,icount,cycles
-		UML_MAPVAR(block, MAPVAR_CYCLES, 0);                                        // mapvar  cycles,0
-		if (allow_exception)
-			UML_EXHc(block, COND_S, *m_out_of_cycles, param);
-																					// exh     out_of_cycles,nextpc
-	}
-	compiler->cycles = 0;
-}
 
 /*-------------------------------------------------
     generate_checksum_block - generate code to
@@ -2430,174 +2306,6 @@ bool sh2_device::generate_group_0(drcuml_block *block, compiler_state *compiler,
 	return false;
 }
 
-bool sh2_device::generate_group_2(drcuml_block *block, compiler_state *compiler, const opcode_desc *desc, uint16_t opcode, int in_delay_slot, uint32_t ovrpc)
-{
-	switch (opcode & 15)
-	{
-	case  0: // MOVBS(Rm, Rn);
-		UML_MOV(block, I0, R32(Rn));        // mov r0, Rn
-		UML_AND(block, I1, R32(Rm), 0xff);  // and r1, Rm, 0xff
-		UML_CALLH(block, *m_write8);
-
-		if (!in_delay_slot)
-			generate_update_cycles(block, compiler, desc->pc + 2, true);
-		return true;
-
-	case  1: // MOVWS(Rm, Rn);
-		UML_MOV(block, I0, R32(Rn));        // mov r0, Rn
-		UML_AND(block, I1, R32(Rm), 0xffff);    // and r1, Rm, 0xffff
-		UML_CALLH(block, *m_write16);
-
-		if (!in_delay_slot)
-			generate_update_cycles(block, compiler, desc->pc + 2, true);
-		return true;
-
-	case  2: // MOVLS(Rm, Rn);
-		UML_MOV(block, I0, R32(Rn));        // mov r0, Rn
-		UML_MOV(block, I1, R32(Rm));        // mov r1, Rm
-		UML_CALLH(block, *m_write32);
-
-		if (!in_delay_slot)
-			generate_update_cycles(block, compiler, desc->pc + 2, true);
-		return true;
-
-	case  3:
-		return false;
-
-	case  4: // MOVBM(Rm, Rn);
-		UML_MOV(block, I1, R32(Rm));        // mov r1, Rm
-		UML_SUB(block, R32(Rn), R32(Rn), 1);    // sub Rn, Rn, 1
-		UML_MOV(block, I0, R32(Rn));        // mov r0, Rn
-		UML_CALLH(block, *m_write8);         // call write8
-
-		if (!in_delay_slot)
-			generate_update_cycles(block, compiler, desc->pc + 2, true);
-		return true;
-
-	case  5: // MOVWM(Rm, Rn);
-		UML_MOV(block, I1, R32(Rm));        // mov r1, Rm
-		UML_SUB(block, R32(Rn), R32(Rn), 2);    // sub Rn, Rn, 2
-		UML_MOV(block, I0, R32(Rn));        // mov r0, Rn
-		UML_CALLH(block, *m_write16);            // call write16
-
-		if (!in_delay_slot)
-			generate_update_cycles(block, compiler, desc->pc + 2, true);
-		return true;
-
-	case  6: // MOVLM(Rm, Rn);
-		UML_MOV(block, I1, R32(Rm));        // mov r1, Rm
-		UML_SUB(block, R32(Rn), R32(Rn), 4);    // sub Rn, Rn, 4
-		UML_MOV(block, I0, R32(Rn));        // mov r0, Rn
-		UML_CALLH(block, *m_write32);            // call write32
-
-		if (!in_delay_slot)
-			generate_update_cycles(block, compiler, desc->pc + 2, true);
-		return true;
-
-	case 13: // XTRCT(Rm, Rn);
-		UML_SHL(block, I0, R32(Rm), 16);        // shl r0, Rm, #16
-		UML_AND(block, I0, I0, 0xffff0000); // and r0, r0, #0xffff0000
-
-		UML_SHR(block, I1, R32(Rn), 16);        // shr, r1, Rn, #16
-		UML_AND(block, I1, I1, 0xffff);     // and r1, r1, #0x0000ffff
-
-		UML_OR(block, R32(Rn), I0, I1);     // or Rn, r0, r1
-		return true;
-
-	case  7: // DIV0S(Rm, Rn);
-		UML_MOV(block, I0, mem(&m_sh2_state->sr));              // move r0, sr
-		UML_AND(block, I0, I0, ~(Q|M|T));       // and r0, r0, ~(Q|M|T) (clear the Q,M, and T bits)
-
-		UML_TEST(block, R32(Rn), 0x80000000);           // test Rn, #0x80000000
-		UML_JMPc(block, COND_Z, compiler->labelnum);            // jz labelnum
-
-		UML_OR(block, I0, I0, Q);               // or r0, r0, Q
-		UML_LABEL(block, compiler->labelnum++);             // labelnum:
-
-		UML_TEST(block, R32(Rm), 0x80000000);           // test Rm, #0x80000000
-		UML_JMPc(block, COND_Z, compiler->labelnum);            // jz labelnum
-
-		UML_OR(block, I0, I0, M);               // or r0, r0, M
-		UML_LABEL(block, compiler->labelnum++);             // labelnum:
-
-		UML_XOR(block, I1, R32(Rn), R32(Rm));           // xor r1, Rn, Rm
-		UML_TEST(block, I1, 0x80000000);            // test r1, #0x80000000
-		UML_JMPc(block, COND_Z, compiler->labelnum);            // jz labelnum
-
-		UML_OR(block, I0, I0, T);               // or r0, r0, T
-		UML_LABEL(block, compiler->labelnum++);             // labelnum:
-		UML_MOV(block, mem(&m_sh2_state->sr), I0);              // mov sr, r0
-		return true;
-
-	case  8: // TST(Rm, Rn);
-		UML_AND(block, I0, mem(&m_sh2_state->sr), ~T);  // and r0, sr, ~T (clear the T bit)
-		UML_TEST(block, R32(Rm), R32(Rn));      // test Rm, Rn
-		UML_JMPc(block, COND_NZ, compiler->labelnum);   // jnz compiler->labelnum
-
-		UML_OR(block, I0, I0, T);   // or r0, r0, T
-		UML_LABEL(block, compiler->labelnum++);         // desc->pc:
-
-		UML_MOV(block, mem(&m_sh2_state->sr), I0);      // mov m_sh2_state->sr, r0
-		return true;
-
-	case 12: // CMPSTR(Rm, Rn);
-		UML_XOR(block, I0, R32(Rn), R32(Rm));   // xor r0, Rn, Rm       (temp)
-
-		UML_SHR(block, I1, I0, 24); // shr r1, r0, #24  (HH)
-		UML_AND(block, I1, I1, 0xff);   // and r1, r1, #0xff
-
-		UML_SHR(block, I2, I0, 16); // shr r2, r0, #16  (HL)
-		UML_AND(block, I2, I2, 0xff);   // and r2, r2, #0xff
-
-		UML_SHR(block, I3, I0, 8);  // shr r3, r0, #8   (LH)
-		UML_AND(block, I3, I3, 0xff);   // and r3, r3, #0xff
-
-		UML_AND(block, I7, I0, 0xff);   // and r7, r0, #0xff    (LL)
-
-		UML_AND(block, mem(&m_sh2_state->sr), mem(&m_sh2_state->sr), ~T);   // and sr, sr, ~T (clear the T bit)
-
-		UML_CMP(block, I1, 0);      // cmp r1, #0
-		UML_JMPc(block, COND_Z, compiler->labelnum);    // jnz labelnum
-		UML_CMP(block, I2, 0);      // cmp r2, #0
-		UML_JMPc(block, COND_Z, compiler->labelnum);    // jnz labelnum
-		UML_CMP(block, I3, 0);      // cmp r3, #0
-		UML_JMPc(block, COND_Z, compiler->labelnum);    // jnz labelnum
-		UML_CMP(block, I7, 0);      // cmp r7, #0
-		UML_JMPc(block, COND_NZ, compiler->labelnum+1); // jnz labelnum
-
-		UML_LABEL(block, compiler->labelnum++);     // labelnum:
-		UML_OR(block, mem(&m_sh2_state->sr), mem(&m_sh2_state->sr), T); // or sr, sr, T
-
-		UML_LABEL(block, compiler->labelnum++);     // labelnum+1:
-		return true;
-
-	case  9: // AND(Rm, Rn);
-		UML_AND(block, R32(Rn), R32(Rn), R32(Rm));  // and Rn, Rn, Rm
-		return true;
-
-	case 10: // XOR(Rm, Rn);
-		UML_XOR(block, R32(Rn), R32(Rn), R32(Rm));  // xor Rn, Rn, Rm
-		return true;
-
-	case 11: // OR(Rm, Rn);
-		UML_OR(block, R32(Rn), R32(Rn), R32(Rm));   // or Rn, Rn, Rm
-		return true;
-
-	case 14: // MULU(Rm, Rn);
-		UML_AND(block, I0, R32(Rm), 0xffff);                // and r0, Rm, 0xffff
-		UML_AND(block, I1, R32(Rn), 0xffff);                // and r1, Rn, 0xffff
-		UML_MULU(block, mem(&m_sh2_state->macl), mem(&m_sh2_state->ea), I0, I1);    // mulu macl, ea, r0, r1
-		return true;
-
-	case 15: // MULS(Rm, Rn);
-		UML_SEXT(block, I0, R32(Rm), SIZE_WORD);                // sext r0, Rm
-		UML_SEXT(block, I1, R32(Rn), SIZE_WORD);                // sext r1, Rn
-		UML_MULS(block, mem(&m_sh2_state->macl), mem(&m_sh2_state->ea), I0, I1);    // muls macl, ea, r0, r1
-		return true;
-	}
-
-	return false;
-}
 
 bool sh2_device::generate_group_3(drcuml_block *block, compiler_state *compiler, const opcode_desc *desc, uint16_t opcode, uint32_t ovrpc)
 {
